Save user stories

+++ api/interfaces/models/IProblemTemplate.cs
@@ -34,5 +34,5 @@ public interface IProblemTemplate
     public Adjectives MartyrAdjectives { get; }
 
 
-    public string GetCharacterStageContribution(int seed, string characterStage, IGenre genre, IArchetype heroArchetype, IArchetype enemyArchetype, IPrimalStakes primalStakes, IDramaticQuestion dramaticQuestion);
+    public string GetCharacterStageContribution(long seed, string characterStage, IGenre genre, IArchetype heroArchetype, IArchetype enemyArchetype, IPrimalStakes primalStakes, IDramaticQuestion dramaticQuestion);
 }
